% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_button.R
\name{generate_button}
\alias{generate_button}
\title{Generate button that hides callouts based on ID}
\usage{
generate_button(
  colour_when_hidden = "black",
  colour_when_showing = "teal",
  text_when_hidden = "Show tips",
  text_when_showing = "Hide tips",
  callout_id = "hide"
)
}
\arguments{
\item{colour_when_hidden}{String. Colour of the toggle switch when callouts are hidden. Default: "teal"}

\item{colour_when_showing}{String. Colour of the toggle switch when callouts are displayed. Default: "black"}

\item{text_when_hidden}{String. Text to display next to the switch when callouts are hidden. Default: "Show tips"}

\item{text_when_showing}{String. Text to display next to the switch when callouts are displayed Default: "Hide tips"}

\item{callout_id}{String. ID that identifies the callouts. Default: "hide". Needs to be set manually to all callouts that are meant to respond to the toggle switch.}
}
\value{
A character string of html, js and css code which generates the toggle switch. Save this string into an object, then call the object in an inline code of a Quarto document.
}
\description{
Generate button that hides callouts based on ID
}
\examples{
 \dontrun{
 ```{r}
 my_awesome_button <- generate_button()
 ```

 Some text

 `r my_awesome_button`

}
}
