% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidr.R
\name{hidr}
\alias{hidr}
\title{idr model for hierarchical stuctured data}
\usage{
hidr(par0_list, X, nlm_Bound = list(low = c(0.05, 0.05, 0.05, 0.5, 0.5, 0.5,
  0.5, 0.05, 0.05, -0.99, -0.99), up = c(0.95, 0.95, 0.95, 20, 20, 20, 20, 10,
  10, 0.99, 0.99)), nlm_control = list(eval.max = 200, iter.max = 100, trace =
  0, rel.tol = 0.01), out_control = list(verbose = list(basic = F, par = F),
  iterMax = 100, eps_loglik = 1, eps_parVec = rep(0.01, length(par0_vec))))
}
\arguments{
\item{expr1}{A n*p matrix or data frame of normalized gene expression data. The rows correspond samples (n), the columns correspond genes (p).}

\item{expr2}{The second gene expression data, should be in the same format, size as expr1.}

\item{s}{The tuning parameter for matrixes differences estimation, leave it as NULL to automatically select.}

\item{s.seq}{The candidates for s selection.}

\item{alpha}{Prespecified level of false discovery rate. A relatively loose criterion is suggested for determines screening matrix.}

\item{verbose}{Set verbose to TURE to show details of s selection.}
}
\description{
The function measures reproducibility of hierarchical structured data (e.g ChIP-seq data from different labs with multiple replicates).
}
\details{
Please refer \bold{``A statistical framework for measuring replicability of heterogeneous omics data from multiple labs'' by Ranalli, Lyu, and Li
}
}
