% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idr2IDR.R
\name{idr2IDR}
\alias{idr2IDR}
\title{Calculates IDR value based on idr value}
\usage{
idr2IDR(idrv)
}
\arguments{
\item{expr1}{A n*p matrix or data frame of normalized gene expression data. The rows correspond samples (n), the columns correspond genes (p).}

\item{expr2}{The second gene expression data, should be in the same format, size as expr1.}

\item{s}{The tuning parameter for matrixes differences estimation, leave it as NULL to automatically select.}

\item{s.seq}{The candidates for s selection.}

\item{alpha}{Prespecified level of false discovery rate. A relatively loose criterion is suggested for determines screening matrix.}

\item{verbose}{Set verbose to TURE to show details of s selection.}
}
\description{
The function Calculates global irreproducibility rate (IDR) given local irreproducibility rate.
}
\details{
Please refer \bold{Yin et.al (2016). Testing differential networks with applications to the detection of gene-gene interactions. Biometrika(2015),pp. 1-20}
}
